Perhaps the last additions
I've migrated to dwm. It's been a while and it appears stable and
lighter than i3. Plus I'm learning a lot via the dwm source code. It's
unlikely that I'd  back to i3, so this might be the last commit to my i3
config.
However it was, it served me well.